Split functionality from server.cpp
